// packages/cc/src/lib/_Types.d.ts
export declare enum BatteryChargingStatus {
    Discharging = 0,
    Charging = 1,
    Maintaining = 2
}
export declare enum BatteryReplacementStatus {
    No = 0,
    Soon = 1,
    Now = 2
}
export declare enum BatteryCommand {
    Get = 2,
    Report = 3,
    HealthGet = 4,
    HealthReport = 5
}